Merge authen

+++ app/Http/Controllers/Api/Admin/OrderController.php
@@ -10,9 +10,23 @@ use Illuminate\Http\Request;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Exception;
 use Symfony\Component\HttpFoundation\Response;
+use App\Models\Product;
 
 class OrderController extends ApiController
 {
+
+    protected $account;
+
+    /**
+     * CategoryController constructor..
+     *
+     * @param UploadImageService   $uploadImageService   UploadImageService
+     */
+    public function __construct()
+    {
+        $this->account = auth('api')->user();
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -22,11 +36,24 @@ class OrderController extends ApiController
     {
         $perPage = $request->perpage ? $request->perpage : config('paginate.number_orders');
         try {
-            $order = Order::with(['user','coupon:id,code,percents', 'processStatus:id,name', 'paymentMethod:id,name'])
-                ->orderFilter($request)->orderBy('created_at', 'desc')->paginate($perPage);
-            return $this->formatPaginate($order);
+            $orders = Order::with(['user','coupon:id,code,percents', 'processStatus:id,name', 'paymentMethod:id,name', 'orderDetails.product'])
+                ->orderFilter($request)->orderBy('created_at', 'desc')->get();
+            $data = $orders->toArray();
+            array_walk($data, function(&$order, $key) {
+                $total = 0;
+                $orderDetails = collect($order['order_details']);
+                foreach ($orderDetails as $key => $orderDetail) {
+                    $total += $orderDetail['price'] * $orderDetail['quantity'];
+                }
+                $order['total_money'] = $total;
+                unset($order['order_details']);
+            });
+            if ($this->account->can('view', Order::all()->first())) {
+                return $this->paginate(collect($data));
+            } else {
+                return $this->errorResponse(config('define.no_authorization'), Response::HTTP_UNAUTHORIZED);
+            }
         } catch (Exception $ex) {
-            dd($ex->getMessage());
             return $this->errorResponse("Orders can not be show.", Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
@@ -41,8 +68,27 @@ class OrderController extends ApiController
     public function show($id)
     {
         try {
+            $manager = accountLogin();
             $order = Order::with(['user', 'coupon', 'processStatus:id,name', 'orderDetails.product', 'paymentMethod:id,name'])->findOrFail($id);
-            return $this->successResponse($order, Response::HTTP_OK);
+            $total = 0;
+            $orderDetails = $order['orderDetails'];
+            foreach ($orderDetails as $key => $orderDetail) {
+                if (isProviderLogin()) {
+                    if (Product::find($orderDetail->product_id)->shop->manager_id != $manager->id) {
+                        $orderDetails->forget($key);
+                    } else {
+                        $total += $orderDetail->price * $orderDetail->quantity;
+                    }
+                } else {
+                    $total += $orderDetail->price * $orderDetail->quantity;                        
+                }
+            }
+            $order['total_money'] = $total;
+            if ($this->account->can('view', $order)) {
+                return $this->successResponse($order, Response::HTTP_OK);
+            } else {
+                return $this->errorResponse(config('define.no_authorization'), Response::HTTP_UNAUTHORIZED);
+            }
         } catch (ModelNotFoundException $ex) {
             return $this->errorResponse("Không có đơn hành cần tìm", Response::HTTP_NOT_FOUND);
         } catch (Exception $ex) {
@@ -62,8 +108,17 @@ class OrderController extends ApiController
     public function update(UpdateOrderRequest $request, Order $order)
     {
         try {
-            if ($order->processing_status == Order::STATUS_PROCESSING) {
-                $order->processing_status = $request->processing_status;
+            if ($this->account->can('update', $order)) {
+                if ($order->processing_status == Order::STATUS_PROCESSING) {
+                    $order->processing_status = $request->processing_status;
+                }
+                $order->payment_status = $request->payment_status;
+                $order->delivery_time = $request->delivery_time;
+                $order->save();
+    
+                return $this->successResponse("Update order successfully", Response::HTTP_OK);
+            } else {
+                return $this->errorResponse(config('define.no_authorization'), Response::HTTP_UNAUTHORIZED);
             }
             $order->payment_status = $request->payment_status;
 //            $order->delivery_time = $request->delivery_time;
@@ -87,13 +142,17 @@ class OrderController extends ApiController
     public function destroy(Order $order)
     {
         try {
-            $order->orderDetails()->delete();
-            $order->delete();
-            return $this->successResponse("Delete order successfully", Response::HTTP_OK);
+            if ($this->account->can('delete', $order)) {
+                $order->orderDetails()->delete();
+                $order->delete();
+                return $this->successResponse("Delete order successfully", Response::HTTP_OK);
+            } else {
+                return $this->errorResponse(config('define.no_authorization'), Response::HTTP_UNAUTHORIZED);
+            }
+            
         } catch (ModelNotFoundException $ex) {
             return $this->errorResponse("Order not found.", Response::HTTP_NOT_FOUND);
         } catch (Exception $ex) {
-            dd($ex->getMessage());
             return $this->errorResponse("Occour error when show order.", Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
